requirements.txt
django~=4.0.2
pytest~=7.0.1
pytest-django
zappa
django-s3-sqlite
mysql-connector-python
mysqlclient
djangorestframework~=3.13.1
markdown
django-filter
django-import-export
sorl-thumbnail
django-storages
Pillow
boto3
Werkzeug
zappa-django-utils
martor